entity_alignment/requirements.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1623867049735/work
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioredis==1.3.1
asgiref==3.4.1
astor==0.8.1
astroid==1.5.3
async-timeout==3.0.1
attrs==21.2.0
blessings==1.7
cachetools==4.2.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
ConfigSpace==0.4.19
contextvars==2.4
coverage @ file:///tmp/build/80754af9/coverage_1614614208500/work
cycler==0.10.0
Cython @ file:///tmp/build/80754af9/cython_1618435161347/work
dataclasses==0.8
dill==0.3.4
Django==3.2.11
emcee==3.1.1
filelock==3.0.12
ftfy==6.0.3
gast @ file:///tmp/build/80754af9/gast_1625073243884/work
gensim==3.8.3
google-api-core==1.31.0
google-auth==1.33.1
googleapis-common-protos==1.53.0
gpustat==0.6.0
grpcio @ file:///tmp/build/80754af9/grpcio_1614884176509/work
h5py @ file:///tmp/build/80754af9/h5py_1593454121459/work
hiredis==2.0.0
idna==3.2
idna-ssl==1.1.0
immutables==0.15
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1617877310517/work
isodate==0.6.1
isort @ file:///tmp/build/80754af9/isort_1628603791788/work
joblib==1.0.1
JPype1==1.3.0
jsonschema==3.2.0
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1610099771815/work
lazy-object-proxy @ file:///tmp/build/80754af9/lazy-object-proxy_1616526919073/work
Markdown @ file:///tmp/build/80754af9/markdown_1614363833670/work
matching==0.1.1
matplotlib @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-suite_1611858699142/work
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
msgpack==1.0.2
multidict==5.1.0
munkres==1.1.4
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
nvidia-ml-py3==7.352.0
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
openbox==0.7.16
opencensus==0.7.13
opencensus-context==0.1.2
# Editable install with no version control (openea==1.0)
-e /data1/TNNLS-Tackle-Response/Phase-I/OpenEA-master/src
packaging==21.0
pandas==0.24.2
patsy==0.5.2
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1625143759384/work
Platypus-Opt==1.0.4
prometheus-client==0.11.0
protobuf==3.17.3
psutil==5.8.0
py-spy==0.3.7
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pydantic==1.8.2
PyGObject==3.40.1
pylint==1.7.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.1
python-igraph @ file:///home/conda/feedstock_root/build_artifacts/python-igraph_1625253447180/work
python-Levenshtein==0.12.2
pytz==2021.1
PyYAML==5.4.1
ray==1.4.1
rdflib==5.0.0
redis==3.5.3
requests==2.26.0
rsa==4.7.2
scikit-learn==0.24.2
scikit-optimize==0.9.0
scipy @ file:///tmp/build/80754af9/scipy_1597686625380/work
singledispatch @ file:///tmp/build/80754af9/singledispatch_1629321204894/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
smart-open==5.1.0
SPARQLWrapper==1.8.5
sqlparse==0.4.2
statsmodels==0.12.2
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminaltables==3.1.10
texttable @ file:///home/conda/feedstock_root/build_artifacts/texttable_1599404337529/work
threadpoolctl==2.1.0
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094701020/work
tqdm==4.62.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
wrapt==1.12.1
yarl==1.6.3
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
